Update addColumn and updateColumn

+++ controllers/boards/deleteBoard.js
@@ -1,20 +1,25 @@
 import { HttpError } from "../../helpers/index.js";
 import Board from "../../models/board.js";
 import Column from "../../models/column.js";
+import Card from "../../models/card.js";
 
 const deleteBoard = async (req, res) => {
     const { id } = req.params;
 
     const result = await Board.findByIdAndRemove(id);
-    console.log(result)
     if (!result) {
         throw HttpError(404, `Board with id=${id} not found`);
     }
 
-    await Column.deleteMany({ board: id });
+    const deletedColumn = await Column.find({ board: id });
+    if (deletedColumn) {
+            deletedColumn.forEach(async (_id) => {
+                await Card.deleteMany({ column: _id });
+            });
+        await Column.deleteMany({ board: id });
+    }
 
-    // TODO delete cards
-    res.status(200).json( {message: "Board deleted", deletedId: result._id });
+    res.status(200).json({message: "Board deleted", deletedId: result._id });
 };
 
 export default deleteBoard;
